refactor: 패키지 구조 및 첫 실행 화면 수정
