+++ app/src/main/AndroidManifest.xml
@@ -14,12 +14,12 @@
         android:theme="@style/Theme.MZ"
         tools:targetApi="31">
 
-        <!-- ✅ 메인 화면 -->
-        <activity android:name=".MainActivity" android:exported="true" />
-
-        <!-- ✅ 앱 실행 시 시작되는 로그인 화면 -->
+        <!-- 로그인 화면 -->
         <activity android:name=".ui.login.LoginActivity"
-            android:exported="true">
+            android:exported="true"/>
+
+        <!-- 메인 화면 -->
+        <activity android:name=".MainActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
